move console context to lapi context
